+++ samples/embed_text.rb
@@ -72,7 +72,7 @@ end
 Squib::Deck.new(cards: 3) do
 
   embed_text = 'Take 1 :tool: and gain 2 :health:.'
-  text(str: embed_text, font: 'Sans', font_size: [18, 25, 32],
+  text(str: embed_text, font: 'Sans', font_size: [18, 32, 45],
        x: 0, y: 0, width: 180, height: 300, valign: :bottom,
        align: :left, ellipsize: false, justify: false, hint: :cyan) do |embed|
     embed.svg key: ':tool:',   width: 28, height: 28, file: 'spanner.svg'